refactor: переименовать класс ExternalKey в AdminProperties.
