+++ src/main/java/ru/elessarov/workout_notebook_bot/utils/AdminProperties.java
@@ -6,12 +6,12 @@ import org.springframework.stereotype.Component;
 
 @Component
 @Getter
-public class ExternalKey {
+public class AdminProperties {
 
     public static String ADMIN_ID;
 
-    @Value("${constants.admin}")
-    public void setAdminStatic(String admin_id) {
-        ExternalKey.ADMIN_ID = admin_id;
+    @Value("${constants.admin_id}")
+    public void setAdminStatic(String adminId) {
+        AdminProperties.ADMIN_ID = adminId;
     }
 }
